Re-working transactions / is_purchased flag.
> Pulled aggregate/custom queries into respective model classes with
@classmethod.
> Pushed transaction sum onto the db instead of a dumb loop.
> The is_purchased flag needs to go on the Item model, but I am waiting.

+++ gifted/main/routes.py
@@ -97,11 +97,7 @@ def register():
 def event(event_id):
     event = Event.query.get(event_id)
     user = User.query.get(session['user_id'])
-    transactions = Transaction.query.filter_by(event_id=event_id, gifter_id=user.id).all()
-    liability = 0
-    for transaction in transactions:
-        liability = liability + transaction.item.price
-
+    liability = Transaction.get_user_liability(event.id, user.id)
     progress = get_event_progress(event_id)
     return render_template('event.html', event=event, progress=progress, logged_in_user=user,
                            liability="{:.2f}".format(liability))
@@ -126,10 +122,11 @@ def wishlist(event_id, user_id):
     event = Event.query.get(event_id)
     user = User.query.get(user_id)
     items = Item.query.filter_by(event_id=event_id, user_id=user_id).all()
+    transactions = Transaction.query.filter_by(event_id=event_id).all()
     my_transactions = Transaction.query.filter_by(event_id=event_id, gifter_id=session['user_id']).all()
     progress = get_wishlist_progress(event_id, user_id)
     return render_template('wishlist.html', event=event, user=user, wishlist=items, progress=progress,
-                           my_transactions=my_transactions)
+                           transactions=transactions, my_transactions=my_transactions)
 
 
 @main.route('/events/<event_id>/purchases/<user_id>')
@@ -180,11 +177,15 @@ def claim_item(event_id, user_id):
 
 @main.route('/events/<event_id>/wishlists/<user_id>/transactions/<transaction_id>/delete', methods=['POST'])
 def unclaim_item(event_id, user_id, transaction_id):
-    transaction_id = request.form.get('transaction_id')
+    if transaction_id != request.form.get('transaction_id'):
+        flash('Transaction identifiers do not match.', 'warning')
+        return redirect(url_for('main.wishlist', event_id=event_id, user_id=user_id))
     transaction = Transaction.query.get(transaction_id)
     item = transaction.item
-    db.session.delete(transaction)
     item.is_purchased = 0
+    db.session.add(item)
+    db.session.commit()
+    db.session.delete(transaction)
     db.session.commit()
     flash('You unclaimed an item!', 'warning')
     return redirect(url_for('main.wishlist', event_id=event_id, user_id=user_id))
@@ -234,35 +235,22 @@ def is_active(starts_on, ends_on):
 
 def get_event_progress(event_id):
     progress = {}
-    user_total_result = Item.query.with_entities(Item.user_id, func.sum(Item.price).label('total')) \
-        .filter_by(event_id=event_id).group_by(Item.user_id).all()
+    user_total_result = Item.get_wishlist_totals(event_id)
     for row in user_total_result:
         user_id = row[0]
         total = row[1]
-        user_purchased_result = Item.query.with_entities(func.sum(Item.price).label('purchased')) \
-            .filter_by(event_id=event_id, user_id=user_id, is_purchased=1).group_by(Item.user_id).first()
-        if user_purchased_result is None:
-            progress[user_id] = {'purchased': '0', 'total': str(total), 'percent': '0'}
-        else:
-            purchased = user_purchased_result[0]
-            percent = purchased / total * 100
-            progress[user_id] = {'purchased': str(purchased), 'total': str(total), 'percent': str(percent)}
+        purchased = Transaction.get_user_total(event_id, user_id)
+        percent = purchased / total * 100 if total != 0 else 0
+        progress[user_id] = {'purchased': str(purchased), 'total': str(total), 'percent': "{:.2f}".format(percent)}
+
+    print(progress)
     return progress
 
 
 def get_wishlist_progress(event_id, user_id):
-    user_total_result = Item.query.with_entities(Item.user_id, func.sum(Item.price).label('total')) \
-        .filter_by(event_id=event_id, user_id=user_id).group_by(Item.user_id).first()
-    user_purchased_result = Item.query.with_entities(func.sum(Item.price).label('purchased')) \
-        .filter_by(event_id=event_id, user_id=user_id, is_purchased=1).group_by(Item.user_id).first()
-    if user_total_result is None:
-        total = 0
-    else:
-        total = user_total_result.total
-    if user_purchased_result is None:
-        progress = {'purchased': '0', 'total': str(total), 'percent': '0'}
-    else:
-        purchased = user_purchased_result[0]
-        percent = purchased / total * 100
-        progress = {'purchased': str(purchased), 'total': str(total), 'percent': str(percent)}
+    total = Item.get_wishlist_total(event_id, user_id)
+    purchased = Transaction.get_user_total(event_id, user_id)
+    percent = purchased / total * 100 if total != 0 else 0
+    progress = {'purchased': str(purchased), 'total': str(total), 'percent': "{:.2f}".format(percent)}
+    print(progress)
     return progress
